requirements.txt
tensorflow-gpu==2.3.0
streamlit==1.11.0
pillow==9.1.0
opencv-python>=4.0.0
matplotlib==3.5.1
seaborn==0.11.2
pandas==1.3.5
numpy==1.18.5
io